+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/CaProviderTest.java
@@ -5,6 +5,7 @@
 package io.strimzi.operator.cluster.operator.assembly;
 
 import io.strimzi.api.kafka.model.common.CertificateAuthorityBuilder;
+import io.strimzi.api.kafka.model.common.CertificateManagerType;
 import io.strimzi.api.kafka.model.kafka.Kafka;
 import io.strimzi.api.kafka.model.kafka.KafkaBuilder;
 import io.strimzi.api.kafka.model.kafka.listener.GenericKafkaListenerBuilder;
@@ -20,6 +21,7 @@ import java.time.Clock;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class CaProviderTest {
     private static final String NAMESPACE = Reconciliation.DUMMY_RECONCILIATION.namespace();
@@ -60,9 +62,11 @@ public class CaProviderTest {
                 caConfig,
                 KAFKA,
                 null,
+                null,
                 CERT_ISSUER,
                 PASSWORD_GENERATOR,
                 Clock.systemUTC(),
+                null,
                 null,
                 null);
 
@@ -80,12 +84,58 @@ public class CaProviderTest {
                 caConfig,
                 KAFKA,
                 null,
+                null,
                 CERT_ISSUER,
                 PASSWORD_GENERATOR,
                 Clock.systemUTC(),
                 null,
+                null,
                 null);
 
         assertThat(provider, instanceOf(CustomCaProvider.class));
+    }
+
+    @Test
+    public void testCreateProviderReturnsCertManagerCaProviderWhenTypeIsCertManager() {
+        CaConfig caConfig = new CaConfig(new CertificateAuthorityBuilder()
+                .withGenerateCertificateAuthority(false)
+                .withType(CertificateManagerType.CERT_MANAGER_IO)
+                .build(), true);
+
+        CaProvider provider = CaProvider.create(Reconciliation.DUMMY_RECONCILIATION,
+                Ca.CaRole.CLUSTER_CA,
+                caConfig,
+                KAFKA,
+                null,
+                null,
+                CERT_ISSUER,
+                PASSWORD_GENERATOR,
+                Clock.systemUTC(),
+                null,
+                null,
+                null);
+
+        assertThat(provider, instanceOf(CertManagerCaProvider.class));
+    }
+
+    @Test
+    public void testCreateProviderThrowsWhenTypeIsCertManagerAndGenerateCaIsTrue() {
+        CaConfig caConfig = new CaConfig(new CertificateAuthorityBuilder()
+                .withGenerateCertificateAuthority(true)
+                .withType(CertificateManagerType.CERT_MANAGER_IO)
+                .build(), true);
+
+        assertThrows(IllegalArgumentException.class, () -> CaProvider.create(Reconciliation.DUMMY_RECONCILIATION,
+                Ca.CaRole.CLUSTER_CA,
+                caConfig,
+                KAFKA,
+                null,
+                null,
+                CERT_ISSUER,
+                PASSWORD_GENERATOR,
+                Clock.systemUTC(),
+                null,
+                null,
+                null));
     }
 }